@@ -1,33 +1,41 @@
 {
-  "name": "gas-frontend",
-  "version": "0.2.0",
+  "name": "app",
+  "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
+    "serve": "vue-cli-service serve --host 0.0.0.0",
     "build": "vue-cli-service build",
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
-    "axios": "^0.19.2",
-    "core-js": "^3.6.4",
-    "echarts": "^4.7.0",
-    "element-ui": "^2.13.1",
-    "qs": "^6.9.4",
-    "vue": "^2.6.11",
-    "vue-axios": "^2.1.5",
-    "vue-router": "^3.3.4",
-    "vuex": "^3.5.1"
+    "@element-plus/icons": "^0.0.11",
+    "@syncfusion/ej2-vue-maps": "^19.2.44",
+    "axios": "^0.21.1",
+    "clipboard": "^2.0.8",
+    "core-js": "^3.6.5",
+    "d3": "^7.0.0",
+    "d3-dsv": "^3.0.1",
+    "element-plus": "^1.0.2-beta.55",
+    "express": "^4.17.1",
+    "plotly.js": "^2.3.1",
+    "vue": "^3.0.0",
+    "vue-axios": "^3.2.4",
+    "vue-image-zoomer": "^1.2.1",
+    "vue-leaflet": "^0.1.0",
+    "vue-outline": "^4.0.4",
+    "vue-resize-directive": "^1.2.0",
+    "vue-router": "^4.0.10",
+    "vue-sidebar-menu": "^4.8.1"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.3.0",
-    "@vue/cli-plugin-eslint": "~4.3.0",
-    "@vue/cli-plugin-router": "^4.4.6",
-    "@vue/cli-plugin-vuex": "^4.4.6",
-    "@vue/cli-service": "~4.3.0",
+    "@vue-leaflet/vue-leaflet": "^0.6.1",
+    "@vue/cli-plugin-babel": "~4.5.0",
+    "@vue/cli-plugin-eslint": "~4.5.0",
+    "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.7.2",
-    "eslint-plugin-vue": "^6.2.2",
-    "vue-template-compiler": "^2.6.11"
+    "eslint-plugin-vue": "^7.0.0"
   },
   "eslintConfig": {
     "root": true,
@@ -35,7 +43,7 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
+      "plugin:vue/vue3-essential",
       "eslint:recommended"
     ],
     "parserOptions": {
