@@ -1,71 +1,97 @@
-import Vue from 'vue'
-import VueRouter from 'vue-router'
+import { createRouter, createWebHistory } from 'vue-router'
+import Home from "../views/Home.vue"
+import BrowseData from "../views/Browse_data.vue"
+import API from '../views/API'
+import Downloads from '../views/Downloads.vue'
+import Help from '../views/Help.vue'
+import AMP_card from '../views/AMP_card'
+import About from '../views/About'
+import Tools from '../views/Tools'
+import Contact from '../views/Contact'
 
-Vue.use(VueRouter)
 
 const routes = [
     {
         path: '/',
-        name: 'Home',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Home.vue')
+        redirect: 'home'
     },
     {
-        path: '/home',
-        name: 'Home',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Home.vue')
+        path: "/home",
+        name: "Home",
+        component: Home,
+        meta:{
+            title: 'AMPSphere: Home'
+        }
     },
     {
-        path: '/amp',
-        name: 'AMP',
-        component: () => import(/* webpackChunkName: "about" */ '../views/AMP.vue')
+        path: "/browse_data",
+        name: "Browse_data",
+        component: BrowseData,
+        meta:{
+            title: 'AMPSphere: Browse data'
+        }
     },
     {
-        path: '/amp_card',
-        name: 'AMP_Card',
-        component: () => import(/* webpackChunkName: "about" */ '../views/AMP_Card.vue')
+        path: "/amp",
+        name: "AMP",
+        component: AMP_card,
+        meta:{
+            title: 'AMPSphere: AMP'
+        },
     },
     {
-        path: '/family',
-        name: 'Family',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Family.vue')
+        path: '/api',
+        name: 'API',
+        component: API,
+        meta:{
+            title: 'AMPSphere: API'
+        }
     },
     {
-        path: '/environment',
-        name: 'Environment',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Environment.vue')
+        path: '/downloads',
+        name: 'Downloads',
+        component: Downloads,
+        meta:{
+            title: 'AMPSphere: Downloads'
+        }
     },
     {
-        path: '/blastp',
-        name: 'Blastp',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Blastp.vue')
-    },
-    {
-        path: '/hmm',
-        name: 'Hmm',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Hmm.vue')
-    },
-    {
-        path: '/macrel',
-        name: 'Macrel',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Macrel.vue')
-    },
-    {
-        path: '/download',
-        name: 'Download',
-        component: () => import(/* webpackChunkName: "about" */ '../views/Download.vue')
+        path: '/help',
+        name: 'Help',
+        component: Help,
+        meta:{
+            title: 'AMPSphere: Help'
+        }
     },
     {
         path: '/contact',
         name: 'Contact',
-        // route level code-splitting
-        // this generates a separate chunk (about.[hash].js) for this route
-        // which is lazy-loaded when the route is visited.
-        component: () => import(/* webpackChunkName: "about" */ '../views/Contact.vue')
+        component: Contact,
+        meta: {
+            title: 'AMPSphere: Contact'
+        }
+    },
+    {
+        path: '/about',
+        name: 'About',
+        component: About,
+        meta: {
+            title: 'AMPSphere: About'
+        }
+    },
+    {
+        path: '/tools',
+        name: 'Tools',
+        component: Tools,
+        meta:{
+            title: 'AMPSphere: Tools'
+        }
     }
 ]
 
-const router = new VueRouter({
-    routes
+const router = createRouter({
+    history: createWebHistory(),
+    routes: routes,
 })
 
 export default router
