@@ -1,9 +1,6 @@
 module.exports = {
-  publicPath: './',
-  outputDir: 'web',
-  devServer: {
-    disableHostCheck: true,
-    port: 8080,
-    public: '127.0.0.1:8080'
-  }
+    devServer: {
+        host: '0.0.0.0',
+        port: '8000'
+    }
 }
